Fragment Page2 created. DetailProductActivity code removed into Page2Fragment. Jetpack Navigation implemented for DetailProductActivity.
