+++ app/src/main/java/com/example/testapplication/di/components/ApplicationComponent.kt
@@ -1,16 +1,11 @@
 package com.example.testapplication.di.components
 
-import android.app.Application
 import android.content.Context
 import com.example.testapplication.di.modules.DataModule
 import com.example.testapplication.di.modules.DomainModule
 import com.example.testapplication.di.modules.ViewModelModule
 import com.example.testapplication.di.scopes.ApplicationScope
-import com.example.testapplication.presentation.activities.DetailProductActivity
-import com.example.testapplication.presentation.fragments.LogInFragment
-import com.example.testapplication.presentation.fragments.Page1Fragment
-import com.example.testapplication.presentation.fragments.ProductImageFragment
-import com.example.testapplication.presentation.fragments.SignInPageFragment
+import com.example.testapplication.presentation.fragments.*
 import dagger.BindsInstance
 import dagger.Component
 
@@ -26,7 +21,7 @@ interface ApplicationComponent {
 
     fun inject(productImageFragment: ProductImageFragment)
 
-    fun inject(detailProductActivity: DetailProductActivity)
+    fun inject(page2Fragment: Page2Fragment)
 
     @Component.Factory
     interface Factory {
