+++ app/src/main/java/com/example/testapplication/presentation/fragments/ProductImageFragment.kt
@@ -64,4 +64,9 @@ class ProductImageFragment : Fragment() {
                 }
             }
     }
+
+    override fun onDestroy() {
+        _binding = null
+        super.onDestroy()
+    }
 }
