+++ app/src/main/res/navigation/nav_graph_detail_product_info.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/nav_graph_detail_product_info"
+    app:startDestination="@id/page2Fragment">
+
+    <fragment
+        android:id="@+id/page2Fragment"
+        android:name="com.example.testapplication.presentation.fragments.Page2Fragment"
+        android:label="Page2Fragment" />
+</navigation>
